requirements.txt
python-dotenv==0.21.0
modal-client==0.45.872
gradio==3.15
discord.py==2.1.0
requests==2.28.2
